Updating environment to python 3.6, cleaning up cohorts.

+++ features/data/variants.py
@@ -68,7 +68,7 @@ def get_variants_mc3(syn):
 
     muts.Sample = muts.Sample.apply(lambda smp: '-'.join(smp.split('-')[:4]))
     muts.PolyPhen = muts.PolyPhen.apply(
-        lambda phen: (gsub('\)$', '', gsub('^.*\(', '', phen))
+        lambda phen: (np.float(gsub('\)$', '', gsub('^.*\(', '', phen)))
                       if phen != '.' else 0)
         )
